- Remove previous word2vec implementation and replace it with the new one. - Add z_score option in trigger_method. - Parameters field of word2vec can now be declared in the configuration file. - Documentation of word2vec in CONFIG_OUTLIERS.md and CONFIG_PARAMETERS.md in progress.
