+++ app/analyzers/ml_models/word2vec.py
@@ -1,383 +1,397 @@
-import numpy as np
-import tensorflow as tf
+from helpers.singletons import logging
+import re
 import math
-import os
-import json
-import collections
-import datetime
-import time
+import copy
 
-from tensorflow.contrib.tensorboard.plugins import projector
-from helpers.singletons import logging, settings
+from collections import Counter
+from itertools import chain
+
+import torch
+import torch.nn as nn
+from torch.utils.data import Dataset, DataLoader
+from torch.optim import Adam
+from torch.optim.optimizer import Optimizer
+
+from typing import List, Tuple, Dict, Optional
 
 
 class Word2Vec:
-    def __init__(self, name):
-        self.batch_size = 128
-        self.embedding_size = 300  # Dimension of the embedding vector
-        self.skip_window = 16  # How many context words to consider left and right of each target word
-        self.num_sampled = self.batch_size / 2  # Number of negative examples to sample
-        self.num_steps = settings.config.getint("machine_learning", "training_steps")
-        self.save_every = self.num_steps / 100  # Save every N steps
-        self.vocabulary_size = None
-
-        # Debug flag
-        self.use_test_data = settings.config.getboolean("machine_learning", "word2vec_use_test_data")
-
-        # Cached probabilities
-        self.use_cache = settings.config.get("machine_learning", "word2vec_use_cache")
-        self.all_probabilities_cache = dict()
-
-        # Set up logging directory
-        self.name = name
-        self.model_name = self.name + "_word2vec"
-        # important: need '' at end so it's treated as directory!
-        self.models_dir = os.path.join(settings.config.get("machine_learning", "models_directory"), self.model_name, '')
-
-        now = datetime.datetime.now()
-        self.log_dir = os.path.join(self.models_dir, now.strftime("%Y-%m-%d %H:%M"), 'log')
-
-        # File locations
-        self.words_to_indices_filename = os.path.join(self.models_dir, self.model_name + "_words_to_indices.json")
-        self.indices_to_words_filename = os.path.join(self.models_dir, self.model_name + "_indices_to_words.json")
-        self.metadata_filename = os.path.join(self.models_dir, self.model_name + "_metadata.tsv")
-
-        self.meta_graph_dir = os.path.join(self.models_dir, self.model_name)
-        self.meta_graph_filename = os.path.join(self.models_dir, self.model_name + ".meta")
-
-        # Create model directory if it does not exist
-        if not os.path.exists(self.models_dir + "/"):
-            os.makedirs(self.models_dir + "/", exist_ok=True)
-
-    def train_model(self, sentences):
-        sentences, words_to_indices, indices_to_words, words = flatten_and_build_indices(sentences)
-
-        # Global position within sentences array
-        sentence_index = 0
-        vocabulary_size = len(set(words))  # Number of unique words in our vocabulary
-
-        logging.logger.debug("number of training sentences: " + str(len(sentences)))
-        logging.logger.debug("words: " + str(len(words)))
-        logging.logger.debug("vocabulary size: " + str(vocabulary_size))
-
-        graph = tf.Graph()
-
-        with graph.as_default():
-            with tf.name_scope('inputs'):
-                # Placeholders are structures for feeding input values
-                train_inputs = tf.placeholder(tf.int32, shape=[self.batch_size])
-                train_labels = tf.placeholder(tf.int32, shape=[self.batch_size, 1])
-
-            # Ops and variables pinned to the CPU because of missing GPU implementation
-            with tf.device('/cpu:0'):
-                # Define embedding matrix variable
-                # Variables are the parameters of the model that are being optimized
-                with tf.name_scope('embeddings'):
-                    embeddings = tf.Variable(tf.random_uniform([vocabulary_size, self.embedding_size], -1.0, 1.0),
-                                             name="embeddings")
-                    # Take an input vector of integer indices,
-                    # and “look up” these indices in the supplied embeddings tensor.
-                    embed = tf.nn.embedding_lookup(embeddings, train_inputs)
-
-                # Construct the variables for the NCE loss
-                with tf.name_scope('weights'):
-                    nce_weights = tf.Variable(tf.truncated_normal([vocabulary_size, self.embedding_size],
-                                                                  stddev=1.0 / math.sqrt(self.embedding_size)),
-                                              name="weights")
-                with tf.name_scope('biases'):
-                    nce_biases = tf.Variable(tf.zeros([vocabulary_size]), name="biases")
-
-            # Compute the average NCE loss for the batch.
-            # tf.nce_loss automatically draws a new sample of the negative labels each
-            # time we evaluate the loss.
-            with tf.name_scope('loss'):
-                loss = tf.reduce_mean(
-                    tf.nn.nce_loss(
-                        weights=nce_weights,
-                        biases=nce_biases,
-                        labels=train_labels,
-                        inputs=embed,
-                        num_sampled=self.num_sampled,
-                        num_classes=vocabulary_size))
-
-            # Add the loss value as a scalar to summary.
-            tf.summary.scalar('loss', loss)
-
-            # Construct the SGD optimizer using a learning rate of 1.0.
-            with tf.name_scope('optimizer'):
-                optimizer = tf.train.GradientDescentOptimizer(1.0).minimize(loss)
-
-            # Compute the cosine similarity between minibatch examples and all embeddings.
-            norm = tf.sqrt(tf.reduce_sum(tf.square(embeddings), 1, keepdims=True), name="norm")
-            # normalized_embeddings = embeddings / norm
-
-            # Merge all summaries.
-            merged = tf.summary.merge_all()
-
-            # Add variable initializer.
-            init = tf.global_variables_initializer()
-
-            # Add saver
-            # Save only latest model
-            saver = tf.train.Saver(max_to_keep=1, save_relative_paths=True)
-
-        # BEGIN TRAINING
-        logging.logger.info("training word2vec model using " + str(len(sentences)) + " samples")
-        logging.init_ticker(total_steps=self.num_steps, desc=self.model_name + " - training word2vec model")
-
-        with tf.Session(graph=graph) as session:
-            # Open a writer to write summaries.
-            writer = tf.summary.FileWriter(self.log_dir, session.graph)
-
-            # We must initialize all variables before we use them.
-            init.run()
-            logging.logger.debug('Initialized all variables')
-            logging.logger.debug(norm)
-
-            average_loss = 0
-            average_historical_loss = list()
-
-            step = 0
-            while step < self.num_steps:
-                logging.tick()
-
-                batch_inputs, batch_labels, sentence_index = generate_batch(self.batch_size, self.skip_window,
-                                                                            sentences, sentence_index)
-                feed_dict = {train_inputs: batch_inputs, train_labels: batch_labels}
-
-                # Define metadata variable.
-                run_metadata = tf.RunMetadata()
-
-                # We perform one update step by evaluating the optimizer op (including it in the list of returned
-                # values for session.run())
-                # Also, evaluate the merged op to get all summaries from the returned "summary" variable.
-                # Feed metadata variable to session for visualizing the graph in TensorBoard.
-                _, summary, loss_val = session.run([optimizer, merged, loss], feed_dict=feed_dict,
-                                                   run_metadata=run_metadata)
-                average_loss += loss_val
-
-                # Add returned summaries to writer in each step.
-                writer.add_summary(summary, step)
-                # Add metadata to visualize the graph for the last run.
-                if step == (self.num_steps - 1):
-                    writer.add_run_metadata(run_metadata, 'step%d' % step)
-
-                if step % 1000 == 0:
-                    if step > 0:
-                        average_loss /= 1000
-                        average_historical_loss.append(average_loss)
-                    # The average loss is an estimate of the loss over the last 1000 steps.
-                    logging.logger.info('average loss at step ' + str(step) + ': ' + str(average_loss))
-                    average_loss = 0
-
-                    # Check if historical loss is showing signs of improvement
-                    if len(average_historical_loss) >= 10:
-                        if np.std(average_historical_loss[-10:]) < 1:
-                            logging.logger.info("loss seems to have stabilized, stopping training process")
-                            step = self.num_steps - 1
-
-                if step % self.save_every == 0:
-                    saver.save(session, self.meta_graph_dir)
-
-                step = step + 1
-
-            # Save used embeddings together with the model
-            with open(self.words_to_indices_filename, "w") as f:
-                json.dump(words_to_indices, f)
-
-            with open(self.indices_to_words_filename, "w") as f:
-                json.dump(indices_to_words, f)
-
-            # Write corresponding labels for the embeddings.
-            with open(self.metadata_filename, 'w') as f:
-                for i in range(vocabulary_size):
-                    f.write(indices_to_words[i] + '\n')
-
-            # Create a configuration for visualizing embeddings with the labels in TensorBoard.
-            config = projector.ProjectorConfig()
-            embedding_conf = config.embeddings.add()
-            embedding_conf.tensor_name = embeddings.name
-            embedding_conf.metadata_path = self.metadata_filename
-            projector.visualize_embeddings(writer, config)
-
-            writer.close()
-
-    def is_trained(self):
-        try:
-            with open(self.words_to_indices_filename, "r"):
-                return True
-        except FileNotFoundError:
-            return False
-
-    def evaluate_sentences(self, sentences):
-        # Load mapping dicts saved while training the model
-        try:
-            with open(self.words_to_indices_filename, "r") as f:
-                words_to_indices = json.load(f)
-        except FileNotFoundError:
-            logging.logger.warn(self.words_to_indices_filename + " not found, did you train the model before " +
-                                "running it?")
-            return
-
-        try:
-            with open(self.indices_to_words_filename, "r") as f:
-                indices_to_words = json.load(f)
-        except FileNotFoundError:
-            logging.logger.warn(self.indices_to_words_filename + " not found, did you train the model before " +
-                                "running it?")
-            return
-
-        graph = tf.Graph()
-        with graph.as_default():
-            saver = tf.train.import_meta_graph(self.meta_graph_filename)
-
-        # DEBUG
-        self.vocabulary_size = len(indices_to_words)
-
-        with tf.Session(graph=graph) as sess:
-            # appends the network defined in .meta file to the current graph
-            saver.restore(sess, tf.train.latest_checkpoint(self.models_dir))
-
-            embeddings = graph.get_tensor_by_name("embeddings/embeddings:0")
-            norm = graph.get_tensor_by_name("norm:0")
-            normalized_embeddings = embeddings / norm
-
-            final_embeddings = normalized_embeddings.eval()
-
-            weights = graph.get_tensor_by_name("weights/weights:0")
-            biases = graph.get_tensor_by_name("biases/biases:0")
-
-            sentences_probs = list()
-
-            for sentence in sentences:
-                tmp_probs = list()
-                for word in sentence:
-                    if word not in words_to_indices.keys():
-                        logging.logger.debug("word " + word + " not known, skipping.")
-                    else:
-                        if self.use_cache:
-                            if word in self.all_probabilities_cache.keys():
-                                all_probabilities = self.all_probabilities_cache[word]
-                            else:
-                                all_probabilities = tf.nn.softmax(tf.nn.xw_plus_b(tf.expand_dims(
-                                    final_embeddings[words_to_indices[word]], 0), tf.transpose(weights), biases)).eval()
-                                self.all_probabilities_cache[word] = all_probabilities
-                        else:
-                            # For each word: Get the probabilities of all context words
-                            all_probabilities = tf.nn.softmax(tf.nn.xw_plus_b(tf.expand_dims(
-                                final_embeddings[words_to_indices[word]], 0), tf.transpose(weights), biases)).eval()
-
-                        word_probs = all_probabilities[0]
-
-                        for target_word in sentence:
-                            if target_word not in words_to_indices.keys():
-                                logging.logger.debug("word " + target_word + " not known, skipping.")
-                            else:
-                                target_word_prob = word_probs[words_to_indices[target_word]]
-                                tmp_probs.append(target_word_prob)
-
-                                if self.use_test_data:
-                                    logging.logger.info("probability of seeing " + word + " in context of " +
-                                                        target_word + " is " + str(target_word_prob))
-
-                # In case we could not calculate any probability, due to all words being unknown, we return NaN
-                if len(tmp_probs) == 0:
-                    tmp_probs = np.nan
-                else:
-                    tmp_probs = np.mean(tmp_probs)
-
-                if "sftp.exe" in sentence:
-                    logging.logger.debug("Sentence:" + str(sentence) + str(tmp_probs))
-                    time.sleep(5)
-                else:
-                    logging.logger.debug("Sentence:" + str(sentence) + str(tmp_probs))
-
-                sentences_probs.append(tmp_probs)
-            return sentences_probs
-
-
-def print_most_matching_words(probs, target_word, indices_to_words, top_n):
-    top_n_indices = (np.argsort(probs)[-top_n:])
-    top_n_indices = np.flip(top_n_indices, axis=0)
-
-    logging.logger.info("-------------------------------")
-    logging.logger.info("Probability of observing " + target_word + " around other words:")
-    for i in top_n_indices:
-        logging.logger.info(indices_to_words[str(i)] + " - " + str(probs[i]))
-
-
-# Unique words must be mapped to integers and vice versa
-# Function requires a flat list as input
-###
-# Dictionary maps words to integers
-# Reversed_dictionary maps integers to words
-###
-def build_mappings(words):
-    # Sort words by their count
-    sort_by_count = []
-    sort_by_count.extend(collections.Counter(words).most_common())
-    dictionary = dict()
-    for word, _ in sort_by_count:
-        dictionary[word] = len(dictionary)
-    reversed_dictionary = dict(zip(dictionary.values(), dictionary.keys()))
-
-    return dictionary, reversed_dictionary
-
-
-def flatten_and_build_indices(sentences):
-    # Flatten list
-    words = [item for sublist in sentences for item in sublist]
-    words_to_indices, indices_to_words = build_mappings(words)
-
-    logging.logger.info("example sentences: " + str(sentences[:3]))
-
-    sentences = [[words_to_indices[word] for word in sent] for sent in sentences]
-    return sentences, words_to_indices, indices_to_words, words
-
-
-# Get all skip gram pairs of a single sentence
-def get_sentence_skipgrams_build(sentence, skip_window):
-    sentence_targets = []
-    sentence_labels = []
-    for i in range(len(sentence)):  # i = target word
-        context_indices = [j for j in range(max(0, (i - skip_window)), min(len(sentence), i + 1 + skip_window))
-                           if j != i]
-        for context_index in context_indices:
-            sentence_targets.append(sentence[i])
-            sentence_labels.append(sentence[context_index])
-    return sentence_targets, sentence_labels
-
-
-# The sentence_labels that this function returns are a list of lists
-#    each inner list contains all the context words of the target word at the corresponding location in the target list
-def get_sentence_skipgrams_restore(sentence, skip_window):
-    sentence_targets = []
-    sentence_labels = []
-    for i in range(len(sentence)):  # i = target word
-        context_indices = [j for j in range(max(0, (i - skip_window)), min(len(sentence), i + 1 + skip_window))
-                           if j != i]
-        sentence_targets.append(sentence[i])
-        labels = []
-        for context_index in context_indices:
-            labels.append(sentence[context_index])
-        sentence_labels.append(labels)
-    return sentence_targets, sentence_labels
-
-
-def generate_batch(batch_size, skip_window, sentences, sentence_index):
-    batch = []
-    labels = []
-    while len(batch) < batch_size:
-        cur_sentence = sentences[sentence_index]
-        sentence_targets, sentence_labels = get_sentence_skipgrams_build(cur_sentence, skip_window)
-        while len(batch) < batch_size and sentence_targets:
-            batch.append(sentence_targets.pop(0))
-            labels.append([sentence_labels.pop(0)])
-        # When at the end of our sentences, start over
-        sentence_index = (sentence_index + 1) % len(sentences)
-    # Backtrack to avoid skipping words at the end of sentences
-    sentence_index = max(0, (sentence_index - 1))
-    # Model requires numpy arrays
-    batch_out = np.asarray(batch, dtype=np.int32)
-    labels_out = np.asarray(labels, dtype=np.int32)
-    return batch_out, labels_out, sentence_index
+    def __init__(self,
+                 separators: str,
+                 size_window: int,
+                 num_epochs: int,
+                 learning_rate: float,
+                 embedding_size: int,
+                 seed: int = 42) -> None:
+        # Default parameters
+        self.train_batch_size = 16
+        self.eval_batch_size = 64
+        self.device = torch.device("cpu")
+        self.unknown_token = 'UNKNOWN'
+        self.num_unknown_occurrence = 0  # number of time the word considered as unknown appear
+
+        self.separators = separators
+        self.size_window = size_window
+        self.epochs = num_epochs  # TODO add early stopping option
+        self.learning_rate = learning_rate
+        self.embedding_size = embedding_size
+        self.seed = seed
+
+        if self.seed != 0:
+            torch.manual_seed(self.seed)
+
+        self.voc_size: int = 0
+        self.voc_counter: Counter = Counter()  # Count the occurrence of the vocabulary
+        self.id2word: Dict[int, str] = dict()
+        self.word2id: Dict[str, int] = dict()
+
+        self.model: Optional[nn.Module] = None
+
+    def train_model(self, train_data: List[str]) -> List[float]:
+        """
+        Train the model self.model with train_data.
+        It will preprocess train_data, create a DataLoader and train a Word2VecModel model on multiple epochs.
+
+        :param train_data: List of string representing the texts that will train the model.
+        :return total_loss_values: List of loss of each training step.
+        """
+        data_preprocessed = self._data_preprocessing(data=train_data)
+
+        train_dataset = Word2VecDataset(data_preprocessed)
+
+        train_data_loader = DataLoader(dataset=train_dataset,
+                                       batch_size=self.train_batch_size,
+                                       shuffle=True)
+
+        if self.model is None:
+            self.model = Word2VecModel(self.embedding_size, self.voc_size).to(self.device)
+
+        adam_optimizer = Adam(self.model.parameters(), lr=self.learning_rate)
+
+        total_loss_values = list()
+        # TODO early stopping
+        for i in range(self.epochs):
+            logging.logger.debug("Train epoch " + str(i))
+            loss_values = train_loop(train_data_loader, self.model, adam_optimizer, self.device)
+            total_loss_values.extend(loss_values)
+        return total_loss_values
+
+    def eval_model(self, eval_data: List[str], output_raw: bool) -> List[Tuple[int, int, int, int, int, float]]:
+        """
+        Evaluate eval_data with the model self.model.
+        It will preprocess eval_data, create a DataLoader and evaluate the data with the model self.model.
+
+        :param eval_data: List of string representing the texts that will be evaluated.
+        :param output_raw: True to have eval_loop that output the raw values and False to have outputs in form of
+        probabilities.
+        :return: List of tuple. Each element within the tuple represent respectively;
+            - center word index in text.
+            - center word id.
+            - context word index in text.
+            - context word id.
+            - text index in eval_data.
+            - probability/raw output of the context word id given the center word id.
+        """
+        data_preprocessed = self._data_preprocessing(data=eval_data)
+
+        eval_dataset = Word2VecDataset(data_preprocessed)
+
+        eval_data_loader = torch.utils.data.DataLoader(eval_dataset,
+                                                       batch_size=self.eval_batch_size,
+                                                       shuffle=False)
+        eval_outputs = eval_loop(eval_data_loader, self.model, self.device, output_raw)
+
+        return eval_outputs
+
+    def update_vocabulary_counter(self, data: List[str]) -> None:
+        """
+        Tokenize the data and update the vocabulary counter.
+
+        :param data: List of string representing texts.
+        """
+        data_tokenized = self._tokenizer(data)
+        self.voc_counter.update(chain.from_iterable(data_tokenized))
+
+    def prepare_voc(self, max_voc_size: int = 6000, min_voc_occurrence: int = 1) -> None:
+        """
+        Create self.voc_size, self.word2id, self.id2word and self.num_unknown_occurrence.
+        If vocabulary is bigger than max_voc_size, it removes the excedent vocabulary that has the smallest occurrence
+        If the occurrence of one word in the vocabulary is smaller than min_voc_occurrence, it remove it from
+        vocabulary.
+
+        :param max_voc_size: Max size of the vocabulary.
+        :param min_voc_occurrence: Minimum time a word has to appear in the dataset.
+        """
+        tmp_voc_dict = dict(self.voc_counter.most_common())
+
+        num_unknown_occurrence = 0
+        for num_occur in tmp_voc_dict.values():
+            if num_occur < min_voc_occurrence:
+                num_unknown_occurrence += num_occur
+        self.num_unknown_occurrence = num_unknown_occurrence
+        if min_voc_occurrence > 1:
+            tmp_voc_dict = {k: v for k, v in tmp_voc_dict.items() if v >= min_voc_occurrence}
+        tmp_voc_list = list(tmp_voc_dict)
+        tmp_voc_size = len(tmp_voc_list)
+        if tmp_voc_size > max_voc_size:
+            tmp_voc_size = max_voc_size
+
+        voc_list = tmp_voc_list[:tmp_voc_size] + [self.unknown_token]
+        self.voc_size = tmp_voc_size + 1
+        self.word2id = {w: idx for (idx, w) in enumerate(voc_list)}
+        self.id2word = {idx: w for (idx, w) in enumerate(voc_list)}
+
+    def _data_preprocessing(self, data: List[str]) -> List[Tuple[int, int, int, int, int]]:
+        """
+        Preprocess data.
+        Tokenize data then convert the tokenized data for word2vec model input.
+
+        :param data: list of string representing texts.
+        :return: List of tuple. Each element within the tuple represent respectively;
+            - center word index in text.
+            - center word id in vocabulary.
+            - context word index in text.
+            - context word id in vocabulary.
+            - text index in eval_data.
+        """
+        data_tokenized = self._tokenizer(data)
+        model_inputs = self._tokenized_texts_to_model_inputs(data_tokenized)
+        return model_inputs
+
+    def _tokenizer(self, data: List[str]) -> List[List[str]]:
+        """
+        Tokenize the data by separating words in texts by the self.separators value.
+
+        :param data: List of string representing texts.
+        :return: List of list of string representing lists of words in a list of texts.
+        """
+        tokens = [re.split(self.separators, x) for x in data]
+        return tokens
+
+    def _tokenized_texts_to_model_inputs(self, tokenized_texts: List[List[str]]) -> List[Tuple[int, int, int, int, int]]:
+        """
+        Convert tokenized text to word2vec model input format.
+
+        :param tokenized_texts: List of list of string representing lists of words in a list of texts.
+        :return: List of tuple. Each element within the tuple represent respectively;
+            - center word index in text.
+            - center word id in vocabulary.
+            - context word index in text.
+            - context word id in vocabulary.
+            - text index in eval_data.
+        """
+        model_inputs = list()
+        for text_idx, text in enumerate(tokenized_texts):
+            for center_idx, center_word in enumerate(text):
+                center_id = self._get_word_id(center_word)
+                first_context_word_index = max(0, center_idx - self.size_window)
+                last_context_word_index = min(center_idx + self.size_window + 1, len(text))
+                for context_idx in range(first_context_word_index, last_context_word_index):
+                    if center_idx != context_idx:
+                        context_word = text[context_idx]
+                        context_id = self._get_word_id(context_word)
+                        model_inputs.append((center_idx, center_id, context_idx, context_id, text_idx))
+        return model_inputs
+
+    def _get_word_id(self, word: str):
+        """
+        Get word id from word.
+
+        :param word: string representing a word.
+        :return: vocabulary id corresponding to the word.
+        """
+        if word in self.word2id:
+            return self.word2id[word]
+        else:
+            return self.word2id[self.unknown_token]
+
+    def prob_model(self, train_data: List[str], output_log_prob: bool) -> List[Tuple[int, int, int, int, int, float]]:
+        """
+        Compute the true probability of each context word context_idx to appear given the center word center_idx.
+        P(context_id|center_id)=(number of time context_id appears with center_id)/(number of time center_id appears)
+        Return same output format than self.eval_model.
+
+        :param train_data: List of string representing the phrases that will train the model
+        :param output_log_prob: If true, convert probabilities to log of probabilities. Used for doing arithmetic mean
+        instead of geometric mean.
+        :return: List of tuple. Each element within the tuple represent respectively;
+            - center word index in text.
+            - center word id.
+            - context word index in text.
+            - context word id.
+            - text index in eval_data.
+            - probability/raw output of the context word id given the center word id.
+        """
+        model_inputs = self._data_preprocessing(data=train_data)
+
+        # total time center_id appears with context_id
+        center_id_to_context_id_to_count: Dict[int, Dict[int, float]] = dict()
+
+        # total time center_id appears
+        center_to_total_count = dict()
+
+        # Computes total time center_id appears with context_id and total time center_id appears
+        for _, center_id, _, context_id, _ in model_inputs:
+            if center_id not in center_id_to_context_id_to_count:
+                center_id_to_context_id_to_count[center_id] = dict()
+                center_to_total_count[center_id] = 1
+            if context_id not in center_id_to_context_id_to_count[center_id]:
+                center_id_to_context_id_to_count[center_id][context_id] = 1
+            center_id_to_context_id_to_count[center_id][context_id] += 1
+            center_to_total_count[center_id] += 1
+
+        # Computes probability context_id appear given center_id
+        center_id_to_context_id_to_prob = copy.deepcopy(center_id_to_context_id_to_count)
+        for center_id, context_count in center_id_to_context_id_to_count.items():
+            for context_id, count in context_count.items():
+                center_id_to_context_id_to_prob[center_id][context_id] = count/center_to_total_count[center_id]
+
+        # Create and return eval_outputs
+        eval_outputs: List[Tuple[int, int, int, int, int, float]] = list()
+        for center_idx, center_id, context_idx, context_id, text_idx in model_inputs:
+            prob = center_id_to_context_id_to_prob[center_id][context_id]
+            if output_log_prob:
+                prob = math.log(prob)
+            eval_outputs.append((center_idx, center_id, context_idx, context_id, text_idx, prob))
+        return eval_outputs
+
+
+class Word2VecModel(nn.Module):
+    """Custom Word2Vec pytorch model."""
+
+    def __init__(self, embedding_size: int, vocab_size: int):
+        """
+        Instantiates one nn.Embedding module and one nn.Linear modules and assign them as member variables.
+
+        :param embedding_size: Size of the embedding vectors.
+        :param vocab_size: Size of the vocabulary.
+        """
+        super(Word2VecModel, self).__init__()
+        self.embeddings = nn.Embedding(vocab_size, embedding_size)
+        self.linear = nn.Linear(embedding_size, vocab_size)
+
+    def forward(self, center_id=None):
+        """
+        Defines the computation.
+
+        :param center_id: Tensor of size [batch_size] representing center_id.
+        :return: raw tensor vector of size [batch_size][vocab_size] representing the context_id.
+        """
+        emb = self.embeddings(center_id)
+        out = self.linear(emb)
+        return out
+
+
+class Word2VecDataset(Dataset):
+    """Custom Word2Vec pytorch Dataset."""
+
+    def __init__(self, model_inputs):
+        self.model_inputs = model_inputs
+
+    def __len__(self):
+        return len(self.model_inputs)
+
+    def __getitem__(self, item):
+        center_idx, center_id, context_idx, context_id, text_idx = self.model_inputs[item]
+        return {"center_idx": center_idx,
+                "center_id": center_id,
+                "context_idx": context_idx,
+                "context_id": context_id,
+                "text_idx": text_idx}
+
+
+def loss_fn(outputs, targets):
+    """Custom pytorch loss function."""
+
+    return nn.CrossEntropyLoss()(outputs, targets)
+
+
+def train_loop(data_loader: DataLoader, model: nn.Module, optimizer: Optimizer, device: torch.device) -> List[float]:
+    """
+    Train loop.
+    Loop in model over input batches, compute loss and do back-propagation.
+
+    :param data_loader: Pytorch DataLoader containing word2vec model inputs.
+    :param model: Word2Vec pytorch model.
+    :param optimizer: Pytorch Optimizer.
+    :param device:
+    :return: List of loss of each training step.
+    """
+    loss_values = list()
+    model.train()
+    for bi, d in enumerate(data_loader):
+        center_id = d["center_id"]
+        context_id = d["context_id"]
+
+        center_id = center_id.to(device, dtype=torch.long)
+        context_id = context_id.to(device, dtype=torch.long)
+
+        optimizer.zero_grad()
+
+        outputs = model(center_id=center_id)
+
+        loss = loss_fn(outputs, context_id)
+        loss_values.append(loss.item())
+
+        loss.backward()
+
+        optimizer.step()
+
+    return loss_values
+
+
+def eval_loop(data_loader: DataLoader,
+              model: nn.Module,
+              device: torch.device,
+              output_raw: bool) -> List[Tuple[int, int, int, int, int, float]]:
+    """
+    Evaluation loop.
+    Loop in model over batches.
+
+    :param data_loader: Pytorch DataLoader containing word2vec model inputs.
+    :param model: Word2Vec pytorch model.
+    :param device:
+    :param output_raw: True to have eval_loop that output the raw values and False to have outputs in form of
+    probabilities.
+    :return: List of tuple. Each element within the tuple represent respectively;
+        - center word index in text.
+        - center word id.
+        - context word index in text.
+        - context word id.
+        - text index in eval_data.
+        - probability/raw output of the context word id given the center word id.
+    """
+    list_center_idx = list()
+    list_center_id = list()
+    list_context_idx = list()
+    list_context_id = list()
+    list_text_idx = list()
+    list_output_val = list()
+    model.eval()
+
+    for bi, d in enumerate(data_loader):
+        center_idx = d["center_idx"]
+        center_id = d["center_id"]
+        context_idx = d["context_idx"]
+        context_id = d["context_id"]
+        text_idx = d["text_idx"]
+
+        center_id = center_id.to(device, dtype=torch.long)
+        context_id = context_id.to(device, dtype=torch.long)
+
+        outputs = model(center_id=center_id)
+        if not output_raw:
+            outputs = nn.Softmax(dim=1)(outputs)  # shape: [batch_size, vocab_size]
+        output_context = outputs[torch.arange(outputs.size(0)), context_id]  # shape: [batch_size]
+
+        list_center_idx.extend(center_idx.tolist())
+        list_center_id.extend(center_id.tolist())
+
+        list_context_idx.extend(context_idx.tolist())
+        list_context_id.extend(context_id.tolist())
+
+        list_text_idx.extend(text_idx.tolist())
+
+        list_output_val.extend(output_context.tolist())
+
+    return list(zip(list_center_idx, list_center_id, list_context_idx, list_context_id, list_text_idx, list_output_val))
