+++ defaults/outliers.conf
@@ -33,10 +33,8 @@ es_wipe_all_existing_outliers=0
 es_wipe_all_whitelisted_outliers=1
 
 # General flags to run, test or train models
-# Training is only used by the Machine Learning models (word2vec)
 run_models=1
 test_models=0
-train_models=0
 
 # 0 for no progress info,  1-4 for progressively more output, 5+ for all the log output
 log_verbosity=1
