--- use_cases/examples_ml/word2vec_dev_suspicious_process_name.conf
@@ -1,46 +1,0 @@
-##############################
-# WORD2VEC - SUSPICIOUS PROCESS NAME
-##############################
-[word2vec_dev_suspicious_process_name]
-es_query_filter=concat: all
-target_fields=word
-aggregator_fields=concat
-
-use_prob_model=0
-use_geo_mean=0
-# use regex format
-separators=' '
-# Size_window has an influence on the memory size
-size_window=3
-# Drop_duplicates has an influence on the memory size because the size of the vocabulary will drop
-drop_duplicates=0
-num_epochs=4
-learning_rate=0.001
-# Embedding_size has en influence on the memory size
-embedding_size=50
-# set seed=0 to make word2vec non deterministic
-seed=42
-# Reduce the memory size has it reduce the size of the vocabulary
-min_uniq_word_occurrence=4
-
-tensorboard=0
-# min_target_buckets has an influence on the memory size
-min_target_buckets = 1000
-
-# word, text
-trigger_focus=word
-# center, context, total or mean (mean compatible only with tirgger_focus=text)
-trigger_score=total
-# low or high --> has to be always low
-trigger_on=low
-#
-trigger_method=stdev
-trigger_sensitivity=5
-
-outlier_type=suspicious process name, attack.t1020, attack.t1041, attack.t1043, attack.t1071
-outlier_reason=test word2vec outlier reason
-outlier_summary=test word2vec outlier summary
-
-run_model=1
-test_model=0
-train_model=1
