+++ use_cases/examples_ml/word2vec_suspicious_process_name.conf
@@ -1,0 +1,21 @@
+##############################
+# WORD2VEC - SUSPICIOUS PROCESS NAME
+##############################
+[word2vec_suspicious_process_name]
+es_query_filter=concat: all
+target=word
+aggregator=concat
+
+#es_query_filter=host: *
+#target_fields=Image
+#aggregator_fields=host
+# use regex format
+#separators='\\'
+
+
+outlier_type=suspicious process name, attack.t1020, attack.t1041, attack.t1043, attack.t1071
+outlier_reason=test word2vec outlier reason
+outlier_summary=test word2vec outlier summary
+
+run_model=1
+test_model=0
